+++ CHANGES.md
@@ -1,4 +1,10 @@
-﻿### 0.13.2 (2017-12-08 19:00:00 UTC)
+﻿### 0.13.3 (2017-12-10 20:30:00 UTC)
+
+* Fix metadata Season Posters and Banners
+* Change restore fetching metadata episode thumbs
+
+
+### 0.13.2 (2017-12-08 19:00:00 UTC)
 
 * Fix tools menu on Chrome mobile browser